Refactor asset and case to async calls, update test to support changes

+++ backend/tests/Services/SubstructureServiceShould.cs
@@ -1,11 +1,9 @@
-
 using api.Adapters;
 using api.Models;
 using api.SampleData.Builders;
 using api.Services;
 
 using Xunit;
-
 
 namespace tests;
 
@@ -59,18 +57,22 @@ public class SubstructureServiceShould : IDisposable
         var expectedSubstructure = CreateTestSubstructure(project);
 
         // Act
-        var projectResult = substructureService.CreateSubstructure(expectedSubstructure, caseId);
+        var projectResult = substructureService.CreateSubstructure(expectedSubstructure, caseId).Result;
 
         // Assert
-        var actualSubstructure = projectResult.Substructures.FirstOrDefault(o => o.Name == expectedSubstructure.Name);
-        Assert.NotNull(actualSubstructure);
-        TestHelper.CompareSubstructures(expectedSubstructure, actualSubstructure);
-        var case_ = fixture.context.Cases.FirstOrDefault(o => o.Id == caseId);
-        Assert.Equal(actualSubstructure.Id, case_.SubstructureLink);
+        if (projectResult.Substructures != null)
+        {
+            var actualSubstructure =
+                projectResult.Substructures.FirstOrDefault(o => o.Name == expectedSubstructure.Name);
+            Assert.NotNull(actualSubstructure);
+            TestHelper.CompareSubstructures(expectedSubstructure, actualSubstructure);
+            var case_ = fixture.context.Cases.FirstOrDefault(o => o.Id == caseId);
+            Assert.Equal(actualSubstructure.Id, case_.SubstructureLink);
+        }
     }
 
     [Fact]
-    public void ThrowNotInDatabaseExceptionWhenCreatingSubstructureWithBadProjectId()
+    public async Task ThrowNotInDatabaseExceptionWhenCreatingSubstructureWithBadProjectId()
     {
         // Arrange
         var loggerFactory = new LoggerFactory();
@@ -81,11 +83,12 @@ public class SubstructureServiceShould : IDisposable
         var expectedSubstructure = CreateTestSubstructure(new Project { Id = new Guid() });
 
         // Act, assert
-        Assert.Throws<NotFoundInDBException>(() => substructureService.CreateSubstructure(expectedSubstructure, caseId));
+        await Assert.ThrowsAsync<NotFoundInDBException>(() =>
+            substructureService.CreateSubstructure(expectedSubstructure, caseId));
     }
 
     [Fact]
-    public void ThrowNotFoundInDatabaseExceptionWhenCreatingSubstructureWithBadCaseId()
+    public async Task ThrowNotFoundInDatabaseExceptionWhenCreatingSubstructureWithBadCaseId()
     {
         // Arrange
         var loggerFactory = new LoggerFactory();
@@ -95,33 +98,49 @@ public class SubstructureServiceShould : IDisposable
         var expectedSubstructure = CreateTestSubstructure(project);
 
         // Act, assert
-        Assert.Throws<NotFoundInDBException>(() => substructureService.CreateSubstructure(expectedSubstructure, new Guid()));
+        await Assert.ThrowsAsync<NotFoundInDBException>(() =>
+            substructureService.CreateSubstructure(expectedSubstructure, new Guid()));
     }
 
     [Fact]
-    public void DeleteSubstructure()
+    public async Task DeleteSubstructure()
     {
         // Arrange
         var loggerFactory = new LoggerFactory();
         var projectService = new ProjectService(fixture.context, loggerFactory);
         var substructureService = new SubstructureService(fixture.context, projectService, loggerFactory);
-        var project = fixture.context.Projects.FirstOrDefault();
-        var substructureToDelete = CreateTestSubstructure(project);
-        var sourceCaseId = project.Cases.FirstOrDefault().Id;
-        substructureService.CreateSubstructure(substructureToDelete, sourceCaseId);
+        if (fixture.context.Projects != null)
+        {
+            var project = fixture.context.Projects.FirstOrDefault();
+            var substructureToDelete = CreateTestSubstructure(project);
+            if (project?.Cases != null)
+            {
+                var sourceCaseId = project.Cases.FirstOrDefault()!.Id;
+                await substructureService.CreateSubstructure(substructureToDelete, sourceCaseId);
+            }
 
-        // Act
-        var projectResult = substructureService.DeleteSubstructure(substructureToDelete.Id);
+            // Act
+            var projectResult = await substructureService.DeleteSubstructure(substructureToDelete.Id);
 
-        // Assert
-        var actualSubstructure = projectResult.Substructures.FirstOrDefault(o => o.Name == substructureToDelete.Name);
-        Assert.Null(actualSubstructure);
-        var casesWithSubstructureLink = projectResult.Cases.Where(o => o.SubstructureLink == substructureToDelete.Id);
-        Assert.Empty(casesWithSubstructureLink);
+            // Assert
+            if (projectResult.Substructures != null)
+            {
+                var actualSubstructure =
+                    projectResult.Substructures.FirstOrDefault(o => o.Name == substructureToDelete.Name);
+                Assert.Null(actualSubstructure);
+            }
+
+            if (projectResult.Cases != null)
+            {
+                var casesWithSubstructureLink =
+                    projectResult.Cases.Where(o => o.SubstructureLink == substructureToDelete.Id);
+                Assert.Empty(casesWithSubstructureLink);
+            }
+        }
     }
 
     [Fact]
-    public void ThrowArgumentExceptionIfTryingToDeleteNonExistentSubstructure()
+    public async Task ThrowArgumentExceptionIfTryingToDeleteNonExistentSubstructure()
     {
         // Arrange
         var loggerFactory = new LoggerFactory();
@@ -129,7 +148,7 @@ public class SubstructureServiceShould : IDisposable
         var substructureService = new SubstructureService(fixture.context, projectService, loggerFactory);
 
         // Act, assert
-        Assert.Throws<ArgumentException>(() => substructureService.DeleteSubstructure(new Guid()));
+        await Assert.ThrowsAsync<ArgumentException>(() => substructureService.DeleteSubstructure(new Guid()));
     }
 
     [Fact]
@@ -186,10 +205,9 @@ public class SubstructureServiceShould : IDisposable
             {
                 Currency = Currency.USD,
                 StartYear = 1030,
-                Values = new double[] { 23.4, 238.9, 32.3 }
+                Values = new[] { 23.4, 238.9, 32.3 },
             }
             );
-
     }
 
     private static Substructure CreateUpdatedSubstructure(Project project, Substructure oldSubstructure)
@@ -207,9 +225,8 @@ public class SubstructureServiceShould : IDisposable
             {
                 Currency = Currency.USD,
                 StartYear = 2030,
-                Values = new double[] { 23.4, 28.9, 32.3 }
+                Values = new[] { 23.4, 28.9, 32.3 },
             }
             );
-
     }
 }
